api1200/deploymentplan updated apiversion to 1020
