+++ examples/image_streamer/deployment_plan.rb
@@ -21,7 +21,7 @@ oneview_client = {
 i3s_client = {
   url: ENV['I3S_URL'],
   oneview_client: oneview_client,
-  api_version: 1200
+  api_version: 1020
 }
 
 # Create or update the Deployment Plan named 'DeploymentPlan1'
