+++ libraries/resource_providers/image_streamer/api1020.rb
@@ -13,8 +13,8 @@ require_relative 'api1000'
 
 module OneviewCookbook
   module ImageStreamer
-    # Module for Image Streamer API 1200 Resources
-    module API1200
+    # Module for Image Streamer API 1020 Resources
+    module API1020
       # Get resource class that matches the type given
       # @param [String] type Name of the desired class type
       # @param [String] variant Variant is ignored
@@ -22,12 +22,12 @@ module OneviewCookbook
       def self.provider_named(type, _variant = nil)
         new_type = type.to_s.downcase.gsub(/[ -_]/, '') + 'provider'
         constants.each do |c|
-          klass = OneviewCookbook::ImageStreamer::API1200.const_get(c)
+          klass = OneviewCookbook::ImageStreamer::API1020.const_get(c)
           next unless klass.is_a?(Class) && klass < OneviewCookbook::ResourceProvider
           name = klass.name.split('::').last.downcase.delete('_').delete('-')
           return klass if new_type =~ /^#{name}$/
         end
-        raise "The '#{type}' resource does not exist for Image Streamer API version 1200."
+        raise "The '#{type}' resource does not exist for Image Streamer API version 1020."
       end
     end
   end
