+++ libraries/resource_providers/image_streamer/api1020/deployment_plan_provider.rb
@@ -1,7 +1,3 @@
-#
-# Cookbook Name:: image_streamer_test_api1200
-# Attributes:: default
-#
 # (c) Copyright 2020 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,8 +8,13 @@
 # under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
-#
 
-default['oneview']['api_version'] = 1200
-
-default['image_streamer_test']['client'] = { url: 'https://imagestreamer.example.com', token: 'token123', api_version: 1200 }
+module OneviewCookbook
+  module ImageStreamer
+    module API1020
+      # Deployment Plan Provider API1020
+      class DeploymentPlanProvider < ImageStreamer::API1000::DeploymentPlanProvider
+      end
+    end
+  end
+end
