+++ spec/fixtures/cookbooks/image_streamer_test_api1020/metadata.rb
@@ -1,0 +1,7 @@
+name             'image_streamer_test_api1020'
+maintainer       'none'
+license          'All rights reserved'
+description      'A test cookbook for the oneview cookbook ImageStreamer::API1020: module'
+version          '0.1.0'
+
+depends          'oneview'
